requirements/requirements.txt
celery==5.3.6
fastapi==0.110.0
pillow==10.2.0
pre-commit==3.6.2
python-multipart==0.0.9
ruff==0.3.0
redis==5.0.2
uvicorn==0.27.1